Added SeekBarPreference to v7 pref support library
Created a SeekBarPreference widget with customizable layout and
attributes controlling the appearance and behavior of the SeekBar such
as whether it's adjustable.
By default, it contains a title, seekbar, and optional seekbar value.
The preference widget is selectable and clickable as a whole rather than the
SeekBar portion of it.
Also, made a leanback-styled SeekBarPreference layout.
The preference supports min attribute as the range's lower bound which does
not require any min attributes support on the framework side.

Change-Id: I368cbe90c9fcd1adf3f8800f63d65ab9b3222280